refactor(workspace): config generation module + atomic file writes (Phase 4)
workspace.rs becomes workspace/mod.rs; write_backend_config and the
OpenCode/Claude Code/Codex config writers (plus their JSON/TOML entry
builders and hook renderers) move verbatim to workspace/config.rs.

Semantic change: all 17 per-workspace config file writes now go through
write_file_atomic (write sibling .tmp + rename), so a crash or a harness
reading mid-write never observes a torn opencode.json /
settings.local.json / config.toml.
